fix(viewer): simplify config with Trunk proxy, remove debug/release split (#211)
- Use relative paths for all API endpoints (Trunk proxy handles routing)
- Remove cfg(debug_assertions) conditional URL logic
- Absorb room_id_from_url() into current_room_id()
- Narrow room_rev type u64 -> u32
- Switch trpg_uses_polling() to false (SSE-first)
- Add Trunk proxy rules for /api/ and /mcp -> localhost:8935
- Add logs/ to .gitignore

+++ viewer/src/game/http.rs
@@ -101,7 +101,7 @@ pub struct InitialStateBuffer {
 #[derive(Resource, Default)]
 pub struct ActiveTrpgRoom {
     pub room_id: String,
-    pub room_rev: u64,
+    pub room_rev: u32,
 }
 
 fn normalize_room_status(raw: &str) -> String {
